+++ packages/manager/src/features/GlobalNotifications/RegionStatusBanner.tsx
@@ -2,7 +2,7 @@ import { Region } from '@linode/api-v4/lib/regions/types';
 import * as React from 'react';
 import Typography from 'src/components/core/Typography';
 import ExternalLink from 'src/components/ExternalLink';
-import Notice from 'src/components/Notice';
+import { Notice } from 'src/components/Notice/Notice';
 import { useRegionsQuery } from 'src/queries/regions';
 
 export interface Props {