+++ packages/manager/src/features/StackScripts/StackScriptBase/StackScriptBase.styles.ts
@@ -19,7 +19,6 @@ const styles = (theme: Theme) =>
     },
     emptyState: {
       color: theme.palette.text.primary,
-      textAlign: 'center',
     },
     table: {
       backgroundColor: theme.bg.bgPaper,