+++ packages/manager/src/features/linodes/LinodesDetail/LinodeNetworking/IPTransfer.tsx
@@ -1,4 +1,4 @@
-import { assignAddresses, IPRange } from '@linode/api-v4/lib/networking';
+import { IPRange } from '@linode/api-v4/lib/networking';
 import { APIError } from '@linode/api-v4/lib/types';
 import {
   both,
@@ -17,25 +17,27 @@ import { StyledActionPanel } from 'src/components/ActionsPanel/ActionsPanel';
 import Button from 'src/components/Button';
 import { CircleProgress } from 'src/components/CircleProgress';
 import Divider from 'src/components/core/Divider';
-import { makeStyles } from '@mui/styles';
+import { makeStyles } from 'tss-react/mui';
 import { Theme } from '@mui/material/styles';
 import Typography from 'src/components/core/Typography';
 import { Dialog } from 'src/components/Dialog/Dialog';
 import Select, { Item } from 'src/components/EnhancedSelect/Select';
 import Grid from '@mui/material/Unstable_Grid2';
-import Notice from 'src/components/Notice';
+import { Notice } from 'src/components/Notice/Notice';
 import usePrevious from 'src/hooks/usePrevious';
-import { ipv6RangeQueryKey } from 'src/queries/networking';
 import {
-  queryKey as linodesQueryKey,
   useAllLinodesQuery,
+  useLinodeQuery,
 } from 'src/queries/linodes/linodes';
-import { useIpv6RangesQuery } from 'src/queries/networking';
 import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
 import { debounce } from 'throttle-debounce';
-import { useQueryClient } from 'react-query';
+import {
+  useAllIPv6RangesQuery,
+  useAssignAdressesMutation,
+  useLinodeIPsQuery,
+} from 'src/queries/linodes/networking';
 
-const useStyles = makeStyles((theme: Theme) => ({
+const useStyles = makeStyles()((theme: Theme) => ({
   sourceIPWrapper: {
     display: 'flex',
     alignItems: 'center',
@@ -82,11 +84,8 @@ const useStyles = makeStyles((theme: Theme) => ({
 }));
 
 interface Props {
-  linodeID: number;
-  linodeRegion: string;
-  ipAddresses: string[];
+  linodeId: number;
   readOnly?: boolean;
-  refreshIPs: () => Promise<void>[];
   open: boolean;
   onClose: () => void;
 }
@@ -114,8 +113,6 @@ interface Swap extends Move {
   selectedLinodesIPs: string[];
 }
 
-type CombinedProps = Props;
-
 const defaultState = (
   sourceIP: string,
   sourceIPsLinodeID: number
@@ -141,22 +138,25 @@ export const getLinodeIPv6Ranges = (
   );
 };
 
-const LinodeNetworkingIPTransferPanel: React.FC<CombinedProps> = (props) => {
-  const {
-    ipAddresses,
-    linodeID,
-    linodeRegion,
-    open,
-    onClose,
-    readOnly,
-  } = props;
-  const classes = useStyles();
-  const queryClient = useQueryClient();
+const LinodeNetworkingIPTransferPanel = (props: Props) => {
+  const { linodeId, open, onClose, readOnly } = props;
+  const { classes } = useStyles();
+  const { mutateAsync: assignAddresses } = useAssignAdressesMutation();
+
+  const { data: linode } = useLinodeQuery(linodeId, open);
+
+  const { data: _ips } = useLinodeIPsQuery(linodeId);
+
+  const publicIPs = _ips?.ipv4.public.map((i) => i.address) ?? [];
+  const privateIPs = _ips?.ipv4.private.map((i) => i.address) ?? [];
+
+  const ipAddresses = [...publicIPs, ...privateIPs];
+
   const [ips, setIPs] = React.useState<IPRowState>(
-    props.ipAddresses.reduce(
+    ipAddresses.reduce(
       (acc, ip) => ({
         ...acc,
-        [ip]: defaultState(ip, linodeID),
+        [ip]: defaultState(ip, linodeId),
       }),
       {}
     )
@@ -186,7 +186,7 @@ const LinodeNetworkingIPTransferPanel: React.FC<CombinedProps> = (props) => {
   } = useAllLinodesQuery(
     {},
     {
-      region: linodeRegion,
+      region: linode?.region,
       label: { '+contains': searchText ? searchText : undefined },
     },
     open // only run the query if the modal is open
@@ -196,9 +196,9 @@ const LinodeNetworkingIPTransferPanel: React.FC<CombinedProps> = (props) => {
     data: ipv6RangesData,
     isLoading: ipv6RangesLoading,
     error: ipv6RangesError,
-  } = useIpv6RangesQuery();
+  } = useAllIPv6RangesQuery();
 
-  const linodes = (allLinodes ?? []).filter((l) => l.id !== linodeID);
+  const linodes = (allLinodes ?? []).filter((l) => l.id !== linodeId);
 
   const onModeChange = (ip: string) => (e: Item) => {
     const mode = e.value as Mode;
@@ -234,7 +234,7 @@ const LinodeNetworkingIPTransferPanel: React.FC<CombinedProps> = (props) => {
           setSelectedIP(ip, firstLinode.ipv4[0]),
           updateSelectedLinodesIPs(ip, () => {
             const linodeIPv6Ranges = getLinodeIPv6Ranges(
-              ipv6RangesData?.data,
+              ipv6RangesData,
               firstLinode.ipv6
             );
             return [...firstLinode.ipv4, ...linodeIPv6Ranges];
@@ -262,7 +262,7 @@ const LinodeNetworkingIPTransferPanel: React.FC<CombinedProps> = (props) => {
             const linode = linodes.find((l) => l.id === Number(e.value));
             if (linode) {
               const linodeIPv6Ranges = getLinodeIPv6Ranges(
-                ipv6RangesData?.data,
+                ipv6RangesData,
                 linode?.ipv6
               );
               return [...linode.ipv4, ...linodeIPv6Ranges];
@@ -411,12 +411,12 @@ const LinodeNetworkingIPTransferPanel: React.FC<CombinedProps> = (props) => {
     if (!equals(previousIPAddresses, ipAddresses)) {
       setIPs(
         ipAddresses.reduce((acc, ip) => {
-          acc[ip] = defaultState(ip, linodeID);
+          acc[ip] = defaultState(ip, linodeId);
           return acc;
         }, {})
       );
     }
-  }, [ipAddresses, linodeID, previousIPAddresses]);
+  }, [ipAddresses, linodeId, previousIPAddresses]);
 
   const ipRow = (ipState: IPStates) => {
     if (isNoneState(ipState)) {
@@ -449,27 +449,12 @@ const LinodeNetworkingIPTransferPanel: React.FC<CombinedProps> = (props) => {
       return;
     }
 
-    assignAddresses(createRequestData(ips, props.linodeRegion))
+    assignAddresses(createRequestData(ips, linode?.region ?? ''))
       .then(() => {
         // Refresh Linodes in the region in which the changes were made.
-        return Promise.all(props.refreshIPs())
-          .then(() => {
-            setSubmitting(false);
-            setError(undefined);
-            setSuccessMessage('IP transferred successfully.');
-            // get updated route_target for ipv6 ranges
-            queryClient.invalidateQueries(ipv6RangeQueryKey);
-            queryClient.invalidateQueries(`${linodesQueryKey}-all`);
-          })
-          .catch((err) => {
-            setError(
-              getAPIErrorOrDefault(
-                err,
-                'Unable to refresh IPs. Please reload the screen.'
-              )
-            );
-            setSubmitting(false);
-          });
+        setSubmitting(false);
+        setError(undefined);
+        setSuccessMessage('IP transferred successfully.');
       })
       .catch((err) => {
         const apiErrors = getAPIErrorOrDefault(
@@ -489,7 +474,7 @@ const LinodeNetworkingIPTransferPanel: React.FC<CombinedProps> = (props) => {
       ipAddresses.reduce(
         (state, ip) => ({
           ...state,
-          [ip]: defaultState(ip, linodeID),
+          [ip]: defaultState(ip, linodeId),
         }),
         {}
       )